+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/rules/IfElseBracingRuleTest.kt
@@ -1,0 +1,336 @@
+package com.pinterest.ktlint.ruleset.standard.rules
+
+import com.pinterest.ktlint.core.api.editorconfig.CODE_STYLE_PROPERTY
+import com.pinterest.ktlint.core.api.editorconfig.CodeStyleValue
+import com.pinterest.ktlint.core.api.editorconfig.CodeStyleValue.ktlint_official
+import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThatRule
+import com.pinterest.ktlint.test.LintViolation
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.EnumSource
+
+class IfElseBracingRuleTest {
+    private val multiLineIfElseRuleAssertThat = assertThatRule { IfElseBracingRule() }
+
+    @ParameterizedTest(name = "CodeStyleValue: {0}")
+    @EnumSource(
+        value = CodeStyleValue::class,
+        mode = EnumSource.Mode.EXCLUDE,
+        names = ["ktlint_official"],
+    )
+    fun `Given another code style then ktlint_official and IF with inconsistent bracing of the branches`(enumValue: CodeStyleValue) {
+        val code =
+            """
+                fun foo() {
+                    if (true) {
+                        doSomething()
+                    } else doSomethingElse()
+                }
+            """.trimIndent()
+        multiLineIfElseRuleAssertThat(code).hasNoLintViolations()
+    }
+
+    @Nested
+    inner class `Given an IF without ELSE` {
+        @Test
+        fun `Given a then-with-braces and no else-branch then do not reformat`() {
+            val code =
+                """
+                fun foo() {
+                    if (true) {
+                        doSomething()
+                    }
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasNoLintViolations()
+        }
+
+        @Test
+        fun `Given a then-without-braces and no else-branch then do not reformat`() {
+            val code =
+                """
+                fun foo() {
+                    if (true) doSomething()
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasNoLintViolations()
+        }
+    }
+
+    @Nested
+    inner class `Given an IF with ELSE` {
+        @Test
+        fun `Given a then-without-braces and an else-without-braces then do not reformat`() {
+            val code =
+                """
+                fun foo() {
+                    if (true)
+                        doSomething()
+                    else doSomethingElse()
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasNoLintViolations()
+        }
+
+        @Test
+        fun `Given a then-with-braces and an else-without-braces then reformat`() {
+            val code =
+                """
+                fun foo() {
+                    if (true) {
+                        doSomething()
+                    } else doSomethingElse()
+                }
+                """.trimIndent()
+            val formattedCode =
+                """
+                fun foo() {
+                    if (true) {
+                        doSomething()
+                    } else {
+                        doSomethingElse()
+                    }
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolation(4, 12, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces")
+                .isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given a then-without-braces and an else-with-braces then reformat`() {
+            val code =
+                """
+                fun foo() {
+                    if (true) doSomething() else {
+                        doSomethingElse()
+                    }
+                }
+                """.trimIndent()
+            val formattedCode =
+                """
+                fun foo() {
+                    if (true) {
+                        doSomething()
+                    } else {
+                        doSomethingElse()
+                    }
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolation(
+                    2,
+                    15,
+                    "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces",
+                )
+                .isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given a then-with-braces and an else-with-braces then do not reformat`() {
+            val code =
+                """
+                fun foo() {
+                    if (true) {
+                        doSomething()
+                    } else {
+                        doSomethingElse()
+                    }
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasNoLintViolations()
+        }
+    }
+
+    @Nested
+    inner class `Given an IF with ELSE-IF` {
+        private val formattedCode =
+            """
+            fun foo(value: int) {
+                if (value > 0) {
+                    doSomething()
+                } else if (value < 0) {
+                    doSomethingElse()
+                } else {
+                    doSomethingElse2()
+                }
+            }
+            """.trimIndent()
+
+        @Test
+        fun `Given all branches without-braces then do not reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0)
+                        doSomething()
+                    else if (value < 0) doSomethingElse()
+                    else doSomethingElse2()
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasNoLintViolations()
+        }
+
+        @Test
+        fun `Given if { -- } else if -- else -- then reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0) {
+                        doSomething()
+                    } else if (value < 0)
+                        doSomethingElse()
+                    else
+                        doSomethingElse2()
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolations(
+                    LintViolation(5, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces"),
+                    LintViolation(7, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces"),
+                ).isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given if -- else if { -- } else -- then reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0)
+                        doSomething()
+                    else if (value < 0) {
+                        doSomethingElse()
+                    } else
+                        doSomethingElse2()
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolations(
+                    LintViolation(3, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces"),
+                    LintViolation(7, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces"),
+                ).isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given if -- else if -- else { -- } then reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0)
+                        doSomething()
+                    else if (value < 0)
+                        doSomethingElse()
+                    else {
+                        doSomethingElse2()
+                    }
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolations(
+                    LintViolation(3, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces"),
+                    LintViolation(5, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces"),
+                ).isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given if { -- } else if { -- } else -- then reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0) {
+                        doSomething()
+                    } else if (value < 0) {
+                        doSomethingElse()
+                    } else
+                        doSomethingElse2()
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolation(7, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces")
+                .isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given if { -- } else if -- else { -- } then reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0) {
+                        doSomething()
+                    } else if (value < 0)
+                        doSomethingElse()
+                    else {
+                        doSomethingElse2()
+                    }
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolation(5, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces")
+                .isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given if -- else if { -- } else { -- } then reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0)
+                        doSomething()
+                    else if (value < 0) {
+                        doSomethingElse()
+                    } else {
+                        doSomethingElse2()
+                    }
+                }
+                """.trimIndent()
+            @Suppress("ktlint:argument-list-wrapping", "ktlint:max-line-length")
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasLintViolation(3, 9, "All branches of the if statement should be wrapped between braces if at least one branch is wrapped between braces")
+                .isFormattedAs(formattedCode)
+        }
+
+        @Test
+        fun `Given all branches with braces then do not reformat`() {
+            val code =
+                """
+                fun foo(value: int) {
+                    if (value > 0) {
+                        doSomething()
+                    } else if (value < 0) {
+                        doSomethingElse()
+                    } else {
+                        doSomethingElse2()
+                    }
+                }
+                """.trimIndent()
+            multiLineIfElseRuleAssertThat(code)
+                .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
+                .hasNoLintViolations()
+        }
+    }
+}